chore: add components to the playground
