+++ packages/_vue3-migration-test/src/x-modules/scroll/index.ts
@@ -1,3 +1,3 @@
+export { default as TestMainScroll } from './test-main-scroll.vue';
 export { default as TestScroll } from './test-scroll.vue';
-export { default as TestBaseScroll } from './test-base-scroll.vue';
 export * from './x-module';
